python_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pypdf==4.3.1
python-docx==1.1.2
reportlab==4.2.2

# Optional (kept light): if you set OPENAI_API_KEY later, you can wire LangChain prompts.
langchain==0.2.14
langchain-text-splitters==0.2.2
